Move definitions to types file
